refractor screens to view
